Progress on the UI
Teach the UI to upload to a bucket, and then download from a different
one. The UI is currently a single HTML file that includes Javascript and
CSS. Pushing the whole thing to a public Casper bucket is how we deploy
it to the real world.

Bonus: also works on mobile!

Double Bonus: created a short URL http://bit.ly/openworld-cloud-events-function